fix(gateway): make terminal completion callbacks type-safe
Await terminal persistence on both optional session-resolution paths and
express command completion with explicit async cleanup. Keep CLI history
regressions with the existing failure-report tests within the current
Gateway test-root limit.
